[client] [dashboard] [labour_mgmt] [work_types] :: Pass WorkTypes as Prop From main Component rather than fetching individually in each component
